+++ .gitignore
@@ -139,3 +139,5 @@ cython_debug/
 
 # macOS
 .DS_Store
+
+.note*
